todo_app/pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = ""
authors = ["dmitriy <dmitriydmgora@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
environs = "^11.0.0"
psycopg = {extras = ["binary"], version = "^3.2.2"}
pytest = "^8.3.3"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
